Maximum length, phone number, city & country name validation - done.
maxlength attributes have been added to inputtext elements in XHTML
(according to the datamodel's properties). Length also checked in Java
code.
2 new Java classes added to the validation package.
One for validating the phone number's format, and one for validating the
city and country names.
The latter could be renamed for something better (for not a so general
one).